safe mode
- allows you to reflash the device even if you've uploaded a sketch that
  interferes with the normal reflashing process (corrupts memory, turns
  off power to USB, etc)
- Thanks @Wozza365 for giving me this idea.

This is also solvable by shorting the reset pin but @Wozza365 was having
a heck of a time getting that working because the 8 seconds that you're
given to start the upload post reset were too short.  His computer was
not  recognizing the USB device in that amount of time.  This tiny
change would have allowed him to boot into safe mode and then take his
time following the typical reflash process.