Add invert() and allPixelsOn() display functions
invert(true) will invert the display so a 0 bit in the buffer will
turn the corresponding pixel on. invert(false) will set to normal.

allPixelsOn(true) will set all pixels on, ignoring the contents of
the buffer but leaving it unaltered. allPixelsOn(false) will display
the contents of the buffer.