Keep nextFrame() the same as V1.1. Add newFrame()
Function nextFrame() is deprecated but left as it is in V1.1, with the timing
bug, so that existing code using it will continue to run at the same speed.

Function newFrame() has been added, which is nextFrame() with the bug fixed.
It should be used for all new development instead of nextFrame().