Add combination display/clear functions
Added core function paintScreenAndClearImage() which is used by
displayAndClear(). These provide the ability to display the buffer and
and also clear it in one operation.

Because the clearing of each buffer byte is done while waiting for its original
value to be shifted out to the display, the clear operation adds no extra time
compared to just using display().